refactor(game): keep reference solutions as typechecked .ts files instead of strings
